[Fix] Make bash the default shell inside emacs
This is an elpy.

[Problem]
For some weird reason, running unittests inside emacs in a virtual enviroment
caused some errors. The python version used to run the tests were from the
system as opposed to the python version inside the virtual environment.

[Fix]
Setting the default shell version as bash appears to fix the problem.

References:
https://github.com/jorgenschaefer/elpy/issues/1001